Issue resolved.
The script checks weather the file that has to be backed up exists or not . If it doesn't exist , it gives an error saying " backup file not found".
If the backup file exists and destination file also exitst , backup is done.
If the backup file exists and destination file doesn't exist , it creates a destination file with the name same as given parameter and backup is done.

+++ ibackup.sh
@@ -4,14 +4,27 @@ if [ $# -eq 0 ]
 then
 	echo " No arguments supplied "
 
+elif [ -e "$1" ]
+ then 
+	if [ -e "$2" ]
+	then
+		var1=$( tar -cvvf backup0.tar -g snap-i $1) 
+       	        var2=$( gzip backup0.tar )  
+	        var3=$( mv backup0.tar.gz snap-i $2)
+
+	else
+
+        mkdir $2	
+        var1=$( tar -cvvf backup0.tar -g snap-i $1) 
+       	var2=$( gzip backup0.tar )  
+	var3=$( mv backup0.tar.gz snap-i $2) 
+        
+        fi
+       
 else
 
-var1=$( tar -cvvf backup0.tar -g snap-i $1)
-var2=$( gzip backup0.tar )
-var3=$( mv backup0.tar.gz snap-i $2)
-
-
+       	echo "The file that has to be backed up doesen't exist"
 fi
 
 echo $#
-echo $1 
+
